pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "prompt-router"
version = "0.1.0"
description = "A TUI prompt router that classifies and sends prompts to Claude or a local Ollama model"
requires-python = ">=3.11"
dependencies = [
  "anthropic>=0.25.0",
  "ollama>=0.2.0",
  "textual>=0.55.0",
  "python-dotenv>=1.0.0",
]

[project.scripts]
prompt-router = "main:main"

[tool.setuptools]
py-modules = ["main"]

[tool.setuptools.packages.find]
where = ["."]
include = ["router*", "ui*"]
